making the generic part of the kernel buildable for 64 bit

+++ src/main.c
@@ -59,8 +59,8 @@ void switch_to_user_mode()
    // maps 16 pages (64 KB) for the user program
 
    map_pages(get_curr_page_dir(),
-             (uint32_t)user_program_vaddr,
-             (uint32_t)user_program_paddr, 16, true, true);
+             (uintptr_t)user_program_vaddr,
+             (uintptr_t)user_program_paddr, 16, true, true);
 
    printk("pdir entries used = %i\n", debug_count_used_pdir_entries(get_curr_page_dir()));
    debug_dump_used_pdir_entries(get_curr_page_dir());
